refactor: Add new way to handle pagination for posts
This commit refactors the PaginationArrowsForPosts component in order to handle the pagination process for posts. Now it checks for the URL to see if there's an element called 'page' and then uses the value of this element to paginate through the API response via the "page[page].astro" page.

Additionally, both "index.astro" pages and the "page[page].astro" page don't use the parameter "fields=intro" when requesting the blog API, since all blog posts that come in the API response contain the "intro" field as default now.