style: Apply new styles for strong and b elements
This commit introduces new rules in the BaseLayout.astro component regarding b and strong elements. Every of these elements now have a new color: text-sky-600.

The only b and strong elements not included in this rule are the ones that are enclosed within an anchor element.

Additionally, every ZipsCard component used in index.astro (src/pages/zips) is enclosed within a div that determines the height of the card, making every card have the same height and displayed as a grid, instead of using flex.

A new Zip article is added: Stored Procedures